调整头文件引用方式，redischannel的调用方式

+++ logic/requestauth_handler.h
@@ -8,12 +8,12 @@
 #ifndef QUESTAUTH_HANDLER_H_
 #define QUESTAUTH_HANDLER_H_
 
-#include "../../common/common_object.h"
-#include "../../frame/frame_impl.h"
-#include "../../frame/redis_session.h"
-#include "../../include/control_head.h"
-#include "../../include/account_msg.h"
-#include "../../include/msg_head.h"
+#include "common/common_object.h"
+#include "frame/frame_impl.h"
+#include "frame/redis_session.h"
+#include "include/control_head.h"
+#include "include/account_msg.h"
+#include "include/msg_head.h"
 #include <string>
 
 using namespace std;
@@ -31,7 +31,7 @@ class CRequstAuthHandler : public CBaseObject
 		ControlHead			m_stCtlHead;
 		MsgHeadCS			m_stMsgHeadCS;
 		CRequestAuthReq		m_stAuthRegistPhoneReq;
-		uint32_t			m_nAuthCode;
+		int64_t				m_nAuthCode;
 		int64_t				m_nAuthCodeExpireTime;
 	};
 public:
